requirements.txt
asgiref==3.8.1
astroid==3.2.2
black==24.4.2
click==8.1.7
coverage==7.5.1
dill==0.3.8
Django==4.2.13
django-localflavor==4.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pylint==3.2.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-stdnum==1.20
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.11.0
